requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
email_validator==2.1.1
python-dotenv==1.0.1
PyMySQL==1.1.0
openpyxl==3.1.5
Flask-Bcrypt==1.0.1
